style updates, reorganizing routes, removing unnecessary home controller
