+++ app/Http/Controllers/CommentController.php
@@ -9,6 +9,12 @@ use Auth;
 
 class CommentController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth');
+        $this->middleware('teamSubscribed');
+    }
     /**
      * Display a listing of the resource.
      *
