+++ app/Http/Controllers/WelcomeController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Auth;
 
 class WelcomeController extends Controller
 {
@@ -14,5 +15,15 @@ class WelcomeController extends Controller
     public function show()
     {
         return view('welcome.show');
+    }
+
+    public function index()
+    {
+        $this->middleware('auth');
+        $this->middleware('teamSubscribed');
+
+        $user = Auth::user();
+        $currentTeam = $user->currentTeam;
+        return redirect()->route('project.index', [$currentTeam->slug]);
     }
 }
