+++ public/mix-manifest.json
@@ -1,6 +1,6 @@
 {
-    "/js/app.js": "/js/app.js?id=0a5ff808ce46a81886b5",
-    "/css/app.css": "/css/app.css?id=21b0aa45d7dca3beb173",
-    "/css/app-rtl.css": "/css/app-rtl.css?id=c8a9ca28ef955b0ba9d9",
+    "/js/app.js": "/js/app.js?id=6656cfd35ee4d4d2a09b",
+    "/css/app.css": "/css/app.css?id=570d15ed1ebabc1e99fb",
+    "/css/app-rtl.css": "/css/app-rtl.css?id=4781253143d728870c61",
     "/js/sweetalert.min.js": "/js/sweetalert.min.js?id=0068f44b0aa1b83fa767"
 }
