+++ resources/assets/sass/components/_dropdown.scss
@@ -6,6 +6,7 @@
         cursor: pointer;
         display: flex;
         padding: 1rem;
+        transition: all .3s;
 
         img, .material-icons {
             margin-right: .5rem;
@@ -22,10 +23,10 @@
     }
 
     &__menu {
+        @extend .box-shadow--far;
         background: $white;
         border: 1px solid $muted;
         border-radius: 3px;
-        box-shadow: 0px 10px 28px -9px rgba(0,0,0,0.75);
         position: absolute;
         top: 4.25rem;
         right: 0;
