+++ resources/assets/sass/components/_header.scss
@@ -1,6 +1,6 @@
 .app-header {
     align-items: center;
-    background: linear-gradient(60deg, $secondary, $primary);
+    background: $black;
     color: $white;
     display: flex;
     padding: 0 2rem;
