+++ resources/assets/sass/components/_widget.scss
@@ -1,5 +1,5 @@
 .widget {
-    border: 1px solid $muted;
+    box-shadow: 0px 10px 13px -7px $black;
 
     + .widget {
         margin-top: 3rem;
@@ -7,7 +7,7 @@
 
     &__header {
         align-items: center;
-        background: $grey-dark;
+        background: $secondary;
         border-radius: 3px 3px 0 0;
         color: $white;
         display: flex;
@@ -19,18 +19,24 @@
     }
 
     &__title {
-        font-size: 1.4rem;
+        font-size: 1.2rem;
         letter-spacing: 1px;
+        text-transform: uppercase;
     }
 
     &__nav {
-        background: $grey;
-        color: $black;
+        background: darken($secondary, 15%);
         display: flex;
         padding: 0 1rem;
 
         a {
+            color: $white;
             padding: 1rem;
+            transition: all .3s;
+
+            &:hover {
+                background: rgba(255,255,255, .1);
+            }
         }
     }
