+++ resources/assets/sass/utilities/_colors.scss
@@ -1,4 +1,4 @@
-$primary:       #76d6ca;
+$primary:       darken(#76d6ca, 10%);
 $secondary:     #416dac;
 $accent:        #FC4A1A;
 $muted:         #dddddd;
@@ -7,4 +7,4 @@ $white:         #FFFFFF;
 $grey-light:    #cccccc;
 $grey:          #aaaaaa;
 $grey-dark:     #666666;
-$black:         #111111;
+$black:         #222222;
