+++ resources/assets/sass/utilities/_layout.scss
@@ -1,3 +1,7 @@
+[v-cloak] {
+    display: none;
+}
+
 .content {
     display: flex;
     justify-content: space-between;
