+++ resources/assets/sass/utilities/_typography.scss
@@ -1,10 +1,10 @@
-@import url('https://fonts.googleapis.com/css?family=Lato:400,900|Material+Icons');
+@import url('https://fonts.googleapis.com/css?family=Lato:300,400,900|Material+Icons');
 
 body {
     font-family: "Lato", sans-serif;
+    font-weight: 300;
 }
 
-a, a:hover {
+a {
     text-decoration: none;
-    color: $black;
 }
