+++ resources/assets/sass/utilities/main.scss
@@ -3,4 +3,5 @@
 @import 'animations';
 @import 'colors';
 @import 'layout';
+@import 'shadows';
 @import 'typography';
