+++ resources/views/layouts/app.blade.php
@@ -39,7 +39,7 @@
                 @component('components.dropdown')
 
                     @slot('label')
-                        <img class="avatar" src="{{ Auth::user()->currentTeam->photo_url }}" />
+                        <img class="avatar--sm" src="{{ Auth::user()->currentTeam->photo_url }}" />
                         {{ Auth::user()->currentTeam->name }}
                     @endslot
 
@@ -51,7 +51,7 @@
                                 {{ $team->name }}
                             </a>
                         @endforeach
-                        <a href="#" class="dropdown__menu-item">
+                        <a href="/settings/teams" class="dropdown__menu-item">
                             <i class="material-icons">people</i>
                             Create Team
                         </a>
@@ -62,7 +62,7 @@
                 @component('components.dropdown')
 
                     @slot('label')
-                        <img class="avatar" src="{{ Auth::user()->photo_url }}" />
+                        <img class="avatar--sm" src="{{ Auth::user()->photo_url }}" />
                         {{ Auth::user()->name }}
                     @endslot
