+++ resources/views/project/show.blade.php
@@ -1,4 +1,4 @@
-@extends('spark::layouts.app')
+@extends('layouts.app')
 
 @section('content')
